move column header functionality from sheets to _core module
